Refactor (mostly) complete! Moved game starting, loading, "global" info into game.rs Moved UI-related functions and info into ui/ Move key input handling in keys.rs, maybe this should move into ui/ Message log in log.rs. Beginnings of a more high-level system begun, will be worked on in a branch AI-related code into ai/ started a rust-related utilities module in utils.rs Moved object info in object/, with acting objects in actor.rs and others in item.rs main.rs doesn't do much for now, might start up configuration there or in game.rs
