+++ src/utils.rs
@@ -1,0 +1,13 @@
+use std::cmp;
+
+pub fn mut_two<T>(first_index: usize, second_index: usize, items: &mut [T]) ->
+    (&mut T, &mut T) {
+        assert!(first_index != second_index);
+        let split_at_index = cmp::max(first_index, second_index);
+        let (first_slice, second_slice) = items.split_at_mut(split_at_index);
+        if first_index < second_index {
+            (&mut first_slice[first_index], &mut second_slice[0])
+        } else {
+            (&mut second_slice[0], &mut first_slice[second_index])
+        }
+    }
